Add isolation machine restriction to run dh_golang_autopkgtest (#2)
* Run dh_golang_autopkgtest with isolation-machine restriction (LP: #1856083)

The auto generated test by autodep8 does run dh_golang_autopkgtest but
it does not have isolation-machine restriction, only allow-stderr. The
TestDevicesSetAllow and TestDevicesSetDeny unit tests need it to
successfully pass.

* d/t/control: Use commas in Restrictions field of basic-smoke test

The documentation mentions that comma should be used when there are
multiple restrictions. It might not be mandatory since without comma it
is working fine but to be complaint with it and follow most of
definitions in other source packages let's add it.

* Update changelog